+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Make `Display::handle_event()`` return a value indicating whether a new
+  primary cycle has begun.
+
+
 ## tiny-led-matrix 0.1.0 (2019-03-10)
 
 * Initial release